+++ src/list/source/links.ts
@@ -5,7 +5,7 @@ import path from 'path'
 import { ListContext, ListItem } from '../../types'
 import BasicList from '../basic'
 import { DocumentLink, Location } from 'vscode-languageserver-types'
-import Uri from 'vscode-uri'
+import { URI } from 'vscode-uri'
 
 export default class LinksList extends BasicList {
   public defaultAction = 'open'
@@ -17,7 +17,7 @@ export default class LinksList extends BasicList {
 
     this.addAction('open', async item => {
       let { target } = item.data
-      let uri = Uri.parse(target)
+      let uri = URI.parse(target)
       if (uri.scheme.startsWith('http')) {
         await nvim.call('coc#util#open_url', target)
       } else {
@@ -70,6 +70,6 @@ export default class LinksList extends BasicList {
 
 function formatUri(uri: string): string {
   if (!uri.startsWith('file:')) return uri
-  let filepath = Uri.parse(uri).fsPath
+  let filepath = URI.parse(uri).fsPath
   return filepath.startsWith(workspace.cwd) ? path.relative(workspace.cwd, filepath) : filepath
 }